pyproject.toml
[project]
name = "mock-hass-websocket"
version = "0.1.0"
description = "A mock Home Assistant websocket server for testing AppDaemon apps."
authors = [
    {name = "Antigravity", email = "antigravity@google.com"},
]
dependencies = [
    "websockets>=12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer>=0.9.0", # For CLI
]
requires-python = ">=3.10"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
mock-hass = "mock_hass_websocket.main:app"
